Feat:학적삭제구현 및 오류수정

+++ back/src/services/certificateService.js
@@ -49,10 +49,13 @@ class CertificateService {
 
     return certificate;
   }
-  static async deleteIdCertificate({ _id }) {
+  static async deleteIdCertificate({ _id, owner }) {
     let certificate = Certificate.findById(_id);
     if (!certificate) {
       return "해당되는 학적이 없습니다.";
+    }
+    if (certificate.owner != owner) {
+      return "본인의 학적이 아닙니다.";
     }
     certificate = await Certificate.deleteById(_id);
     return certificate;
